requirements.txt
pytest
pytest-bdd
redis
python-dotenv